requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
websockets==15.0.1
yml==0.0.1
